getSmoothADC() changed to type Unsigned Long. getSmoothADC() now returns ADC data on request. Timer init and config added to ADC setup for 50KSPS timer triggering.

+++ include/shared_adc.h
@@ -47,7 +47,7 @@ void initAdcModule(char adcs);
  *
  * \return The smoothed value from the ADC.
  */
-int getSmoothAdc(char adc);
+unsigned long getSmoothAdc(char adc);
 
 #endif
